+++ hw15-MessageSystem/src/main/java/ru/otus/springmvcwebapp/api/hibernatedao/UserDaoException.java
@@ -1,4 +1,4 @@
-package ru.otus.springmvcwebapp.api.dao;
+package ru.otus.springmvcwebapp.api.hibernatedao;
 
 public class UserDaoException extends RuntimeException {
   public UserDaoException(Exception ex) {
