+++ hw15-MessageSystem/src/main/java/ru/otus/springmvcwebapp/appconfig/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package ru.otus.springmvcwebapp.chat;
+package ru.otus.springmvcwebapp.appconfig;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
@@ -18,7 +18,6 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
 
     @Override
     public void registerStompEndpoints(StompEndpointRegistry registry) {
-        registry.addEndpoint("/chat/websocket").withSockJS();
         registry.addEndpoint("/users/showFormForAdd/websocket").withSockJS();
     }
