+++ hw15-MessageSystem/src/main/java/ru/otus/springmvcwebapp/appconfig/security/WebAppSecurityConfig.java
@@ -40,7 +40,6 @@ public class WebAppSecurityConfig extends WebSecurityConfigurerAdapter {
                 .antMatchers(HttpMethod.GET, "/").permitAll()
                 .antMatchers(HttpMethod.GET, "/resources/static/**").permitAll()
                 .antMatchers("/webjars/**").permitAll()
-                .antMatchers("/chat/**").permitAll()
                 .anyRequest().authenticated()
                 .and()
                 .formLogin()
