--- hw15-MessageSystem/src/main/java/ru/otus/springmvcwebapp/chat/ChatMessage.java
@@ -1,14 +1,0 @@
-package ru.otus.springmvcwebapp.chat;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class ChatMessage {
-
-    private String messageStr;
-
-}
