--- hw15-MessageSystem/src/main/java/ru/otus/springmvcwebapp/chat/ChatMessageController.java
@@ -1,19 +1,0 @@
-package ru.otus.springmvcwebapp.chat;
-
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.SendTo;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.util.HtmlUtils;
-
-@Controller
-public class ChatMessageController {
-
-
-    @MessageMapping("/message")
-    @SendTo("/topic/chatMessages")
-    public ChatMessage showMessage(ChatMessage message) throws Exception {
-//        Thread.sleep(1000); // simulated delay
-        return new ChatMessage(HtmlUtils.htmlEscape(message.getMessageStr()));
-    }
-
-}
