+++ hw15-MessageSystem/src/main/java/ru/otus/springmvcwebapp/controllers/AdminPanelController.java
@@ -1,24 +1,15 @@
 package ru.otus.springmvcwebapp.controllers;
 
-import com.google.gson.Gson;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.messaging.core.AbstractDestinationResolvingMessagingTemplate;
 import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import ru.otus.springmvcwebapp.api.services.DBServiceCachedUser;
 import ru.otus.springmvcwebapp.front.FrontendService;
 import ru.otus.springmvcwebapp.repository.User;
-
-import javax.annotation.PostConstruct;
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * @author Sergei Viacheslaev
@@ -34,26 +25,9 @@ public class AdminPanelController {
 
     private SimpMessagingTemplate messageSender;
 
-    public AdminPanelController(DBServiceCachedUser serviceUser) {
+    public AdminPanelController(DBServiceCachedUser serviceUser, FrontendService frontendService, SimpMessagingTemplate messageSender) {
         this.serviceUser = serviceUser;
-    }
-
-    @PostConstruct
-    private void init() {
-        //cache and DB
-        serviceUser.saveUser(new User("Vasya", "Pupkin", 22));
-        serviceUser.saveUser(new User("Tom", "Hanks", 65));
-        serviceUser.saveUser(new User("Bill", "Gates", 51));
-        serviceUser.saveUser(new User("Maulder", "Fox", 35));
-    }
-
-    @Autowired
-    public void setFrontendService(FrontendService frontendService) {
         this.frontendService = frontendService;
-    }
-
-    @Autowired
-    public void setMessageSender(SimpMessagingTemplate messageSender) {
         this.messageSender = messageSender;
     }
