--- hw15-MessageSystem/src/main/java/ru/otus/springmvcwebapp/hibernate/HibernateUtils.java
@@ -1,30 +1,0 @@
-package ru.otus.springmvcwebapp.hibernate;
-
-import org.hibernate.SessionFactory;
-import org.hibernate.boot.Metadata;
-import org.hibernate.boot.MetadataSources;
-import org.hibernate.boot.registry.StandardServiceRegistry;
-import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
-import org.hibernate.cfg.Configuration;
-
-import java.util.Arrays;
-
-public final class HibernateUtils {
-
-  private HibernateUtils() {
-  }
-
-  public static SessionFactory buildSessionFactory(String configResourceFileName, Class...annotatedClasses){
-    Configuration configuration = new Configuration().configure(configResourceFileName);
-    MetadataSources metadataSources = new MetadataSources(createServiceRegistry(configuration));
-    Arrays.stream(annotatedClasses).forEach(metadataSources::addAnnotatedClass);
-
-    Metadata metadata = metadataSources.getMetadataBuilder().build();
-    return metadata.getSessionFactoryBuilder().build();
-  }
-
-  private static StandardServiceRegistry createServiceRegistry(Configuration configuration){
-    return new StandardServiceRegistryBuilder()
-        .applySettings(configuration.getProperties()).build();
-  }
-}
