+++ hw15-MessageSystem/src/main/java/ru/otus/springmvcwebapp/hibernate/dao/UserDaoHibernate.java
@@ -5,8 +5,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
-import ru.otus.springmvcwebapp.api.dao.UserDao;
-import ru.otus.springmvcwebapp.api.dao.UserDaoException;
+import ru.otus.springmvcwebapp.api.hibernatedao.UserDao;
+import ru.otus.springmvcwebapp.api.hibernatedao.UserDaoException;
 import ru.otus.springmvcwebapp.api.sessionmanager.SessionManager;
 import ru.otus.springmvcwebapp.hibernate.sessionmanager.DatabaseSessionHibernate;
 import ru.otus.springmvcwebapp.hibernate.sessionmanager.SessionManagerHibernate;
