+++ hw15-MessageSystem/src/main/webapp/WEB-INF/view/home-page.html
@@ -8,6 +8,5 @@
 <img th:src="@{/resources/static/img/tomcat_logo.png}" width="300" height="150" alt="tomcat_logo"/> <br> <br>
 
 <a th:href="@{login}">Login to server</a> <br>
-<a th:href="@{chat}">Enter chat</a>
 </body>
 </html>
